+++ examples/Basic/containers/ProfileScreen.tsx
@@ -33,8 +33,9 @@ const ProfileScreen: FC<ProfileScreenProps> = ({navigation}) => {
 
   const onLoginPress = useCallback(() => {
     // b909a31c-7a73-11ed-a1eb-0242ac120002
+    // f1851c87-e0ff-4349-a824-cd9b5e5211b9
     NamiCustomerManager.login(
-      'f1851c87-e0ff-4349-a824-cd9b5e5211b9',
+      'E97EDA7D-F1BC-48E1-8DF4-F67EF4A4E4FF',
       (success, error) => {
         setIsUserLogin(success);
         checkId();
