+++ ios/NamiPurchaseManagerBridge.swift
@@ -3,7 +3,7 @@
 //  RNNami
 //
 //  Created by macbook on 06.04.2023.
-//  Copyright © 2023 Facebook. All rights reserved.
+//  Copyright © 2023 Nami ML INc.. All rights reserved.
 //
 
 import Foundation
@@ -11,19 +11,6 @@ import NamiApple
 import React
 
 @objc(RNNamiPurchaseManager)
-class RNNamiPurchaseManager: RCTEventEmitter {
-    override func supportedEvents() -> [String]! {
-      return ["RegisterBuySKU"]
-    }
-    
-    @objc(buySkuComplete:)
-    func buySkuComplete(callback: @escaping RCTResponseSenderBlock) -> Void {
-//        NamiPurchaseManager.buySkuComplete()
-    }
-    
-    @objc(registerBuySkuHandler)
-    func registerBuySkuHandler() {
-
-    }
+class RNNamiPurchaseManager: NSObject {
 }
