+++ src/Nami.d.ts
@@ -1,0 +1,108 @@
+export const Nami: {
+  configure: (config: NamiConfiguration) => void;
+};
+
+export type NamiConfiguration = {
+  appPlatformId: string;
+  logLevel: string;
+  namiLanguageCode?: NamiLanguageCodes;
+  bypassStore?: boolean;
+};
+
+export type NamiLanguageCodes =
+  | "af"
+  | "ar"
+  | "ar-dz"
+  | "ast"
+  | "az"
+  | "bg"
+  | "be"
+  | "bn"
+  | "br"
+  | "bs"
+  | "ca"
+  | "cs"
+  | "cy"
+  | "da"
+  | "de"
+  | "dsb"
+  | "el"
+  | "en"
+  | "en-au"
+  | "en-gb"
+  | "eo"
+  | "es"
+  | "es-ar"
+  | "es-co"
+  | "es-mx"
+  | "es-ni"
+  | "es-ve"
+  | "et"
+  | "eu"
+  | "fa"
+  | "fi"
+  | "fr"
+  | "fy"
+  | "ga"
+  | "gd"
+  | "gl"
+  | "he"
+  | "hi"
+  | "hr"
+  | "hsb"
+  | "hu"
+  | "hy"
+  | "ia"
+  | "id"
+  | "ig"
+  | "io"
+  | "is"
+  | "it"
+  | "ja"
+  | "ka"
+  | "kab"
+  | "kk"
+  | "km"
+  | "kn"
+  | "ko"
+  | "ky"
+  | "lb"
+  | "lt"
+  | "lv"
+  | "mk"
+  | "ml"
+  | "mn"
+  | "mr"
+  | "my"
+  | "nb"
+  | "ne"
+  | "nl"
+  | "nn"
+  | "os"
+  | "pa"
+  | "pl"
+  | "pt"
+  | "pt-br"
+  | "ro"
+  | "ru"
+  | "sk"
+  | "sl"
+  | "sq"
+  | "sr"
+  | "sr-latn"
+  | "sv"
+  | "sw"
+  | "ta"
+  | "te"
+  | "tg"
+  | "th"
+  | "tk"
+  | "tr"
+  | "tt"
+  | "udm"
+  | "uk"
+  | "ur"
+  | "uz"
+  | "vi"
+  | "zh-hans"
+  | "zh-hant";
