+++ src/Nami.js
@@ -1,0 +1,7 @@
+import { NativeModules } from "react-native";
+
+export const { NamiBridge } = NativeModules;
+
+export const Nami = {
+ ...NamiBridge
+};
