+++ src/NamiAnalytics.js
@@ -1,0 +1,7 @@
+import { NativeModules } from "react-native";
+
+export const { NamiAnalyticsEmitter } = NativeModules;
+
+export const NamiAnalytics = {
+ ...NamiAnalyticsEmitter
+};
