+++ src/NamiCampaignManager.js
@@ -1,0 +1,22 @@
+import { NativeModules, NativeEventEmitter } from "react-native";
+
+export const { RNNamiCampaignManager } = NativeModules;
+
+export const NamiCampaignManager = {
+  emitter: new NativeEventEmitter(RNNamiCampaignManager),
+  ...RNNamiCampaignManager,
+  launch: (label, callback) => {
+    RNNamiCampaignManager.launch(label ?? null, callback ?? (() => {}));
+  },
+  isCampaignAvailable: (label) => {
+    return RNNamiCampaignManager.isCampaignAvailable(label ?? null);
+  },
+  registerAvailableCampaignsHandler(callback) {
+    const subscription = this.emitter.addListener(
+      "AvailableCampaignsChanged",
+      callback
+    );
+    RNNamiCampaignManager.registerAvailableCampaignsHandler();
+    return subscription.remove;
+  },
+};
