+++ src/NamiMLManager.d.ts
@@ -1,0 +1,5 @@
+export const NamiMLManager: {
+    coreAction: (label: string) => void;
+    enterCoreContent: (label: string | string[]) => void;
+    exitCoreContent: (label: string | string[]) => void;
+};
