+++ src/NamiMLManager.js
@@ -1,0 +1,7 @@
+import { NativeModules } from "react-native";
+
+export const { NamiMLManagerBridge } = NativeModules;
+
+export const NamiMLManager = {
+ ...NamiMLManagerBridge
+};
